+++ addon/components/groupedit-toolbar.js
@@ -4,6 +4,8 @@
 
 import { inject as service } from '@ember/service';
 import { assert } from '@ember/debug';
+import { computed } from '@ember/object';
+import { isNone } from '@ember/utils';
 import FlexberryBaseComponent from './flexberry-base-component';
 
 /**
@@ -96,6 +98,15 @@ export default FlexberryBaseComponent.extend({
   */
   defaultSettingsButton: true,
 
+  /**
+    Boolean property to show or hide arrows button in toolbar.
+
+    @property arrowsButtons
+  */
+  arrowsButtons: computed('orderedProperty', function() {
+    return !isNone(this.get('orderedProperty'));
+  }),
+
   actions: {
     /**
       Handles add record button click and triggers add record event on
@@ -178,6 +189,20 @@ export default FlexberryBaseComponent.extend({
       } else {
         throw new Error('Unsupported action type for custom buttons.');
       }
+    },
+
+    /**
+      Handles arrow buttons click.
+
+      @method actions.moveRow
+    */
+    moveRow(shift) {
+      if (this.get('readonly')) {
+        return;
+      }
+
+      let componentName = this.get('componentName');
+      this.get('_groupEditEventsService').moveRowTrigger(componentName, shift);
     },
   },