+++ blueprints/flexberry-model/ModelBlueprint.js
@@ -135,17 +135,18 @@ var ModelBlueprint = /** @class */ (function () {
                         TAB + TAB + ("@property " + attr.name + "\n") +
                         TAB + "*/\n" + TAB;
             }
-            var defaultValue = "";
+            var options = [];
+            var optionsStr = "";
             if (attr.defaultValue) {
                 switch (attr.type) {
                     case 'decimal':
                     case 'number':
                     case 'boolean':
-                        defaultValue = ", { defaultValue: " + attr.defaultValue + " }";
+                        options.push("defaultValue: " + attr.defaultValue);
                         break;
                     case 'date':
                         if (attr.defaultValue === 'Now') {
-                            defaultValue = ", { defaultValue() { return new Date(); } }";
+                            options.push("defaultValue() { return new Date(); }");
                             break;
                         }
                         break;
@@ -153,14 +154,28 @@ var ModelBlueprint = /** @class */ (function () {
                         if (this.enums.hasOwnProperty(attr.type)) {
                             var enumName = this.enums[attr.type].className + "Enum";
                             this.enumImports[enumName] = attr.type;
-                            defaultValue = ", { defaultValue: " + enumName + "." + attr.defaultValue + " }";
+                            options.push("defaultValue: " + enumName + "." + attr.defaultValue);
                         }
                         else {
-                            defaultValue = ", { defaultValue: '" + attr.defaultValue + "' }";
+                            options.push("defaultValue: '" + attr.defaultValue + "'");
                         }
                 }
             }
-            attrs.push("" + comment + attr.name + ": DS.attr('" + attr.type + "'" + defaultValue + ")");
+            if (attr.ordered) {
+                options.push("ordered: true");
+            }
+            if (options.length != 0) {
+                optionsStr = ", { " + options.join(', ') + ' }';
+            }
+            attrs.push("" + comment + attr.name + ": DS.attr('" + attr.type + "'" + optionsStr + ")");
+            if (attr.notNull) {
+                if (attr.type === "date") {
+                    validations.push(attr.name + ": { datetime: true }");
+                }
+                else {
+                    validations.push(attr.name + ": { presence: true }");
+                }
+            }
             if (attr.stored)
                 continue;
             var methodToSetNotStoredProperty = "/**\n" +
@@ -237,11 +252,16 @@ var ModelBlueprint = /** @class */ (function () {
         }
         var hiddenStr = "";
         if (belongsTo.hidden || belongsTo.index == -1) {
-            hiddenStr = ", { hidden: true }";
+            hiddenStr = ", { index: " + belongsTo.index + ", hidden: true }";
         }
         else {
-            if (belongsTo.lookupValueField)
-                hiddenStr = ", { displayMemberPath: '" + belongsTo.lookupValueField + "' }";
+            index = belongsToAttrs[0].index;
+            if (belongsTo.lookupValueField) {
+              hiddenStr = ", { index: " + belongsTo.index + ", displayMemberPath: '" + belongsTo.lookupValueField + "' }";
+            }
+            else {
+              hiddenStr = ", { index: " + belongsTo.index + " }";
+            }
         }
         var indent = [];
         for (var i = 0; i < level; i++) {
@@ -266,7 +286,10 @@ var ModelBlueprint = /** @class */ (function () {
     ModelBlueprint.prototype.declareProjAttr = function (attr) {
         var hiddenStr = "";
         if (attr.hidden) {
-            hiddenStr = ", { hidden: true }";
+            hiddenStr = ", { index: " + attr.index + ", hidden: true }";
+        }
+        else {
+            hiddenStr = ", { index: " + attr.index + " }";
         }
         return new SortedPair(attr.index, attr.name + ": attr('" + attr.caption + "'" + hiddenStr + ")");
     };