+++ tests/dummy/app/controllers/components-examples/flexberry-objectlistview/settings-example.js
@@ -175,6 +175,15 @@ export default ListFormController.extend({
   refreshButton: false,
 
   /**
+    Flag: indicates whether 'flexberry-objectlistview' component is in 'defaultSortingButton' mode or not
+
+    @property defaultSortingButton
+    @type Boolean
+    @default true
+  */
+  defaultSortingButton: true,
+
+  /**
     Flag: indicates whether 'flexberry-objectlistview' component is in 'showCheckBoxInRow' mode or not.
 
     @property showCheckBoxInRow
@@ -275,6 +284,14 @@ export default ListFormController.extend({
   availableCollExpandMode: false,
 
   /**
+    Flag for 'flexberry-objectlistview' component 'fixedHeader' property.
+
+    @property fixedHeader
+    @type Boolean
+   */
+  fixedHeader: false,
+
+  /**
     Current records.
 
     @property _records
@@ -317,6 +334,7 @@ export default ListFormController.extend({
       '  applyFilters=(action "applyFilters")<br>' +
       '  resetFilters=(action "resetFilters")<br>' +
       '  refreshButton=refreshButton<br>' +
+      '  defaultSortingButton=defaultSortingButton<br>' +
       '  filterButton=filterButton<br>' +
       '  showCheckBoxInRow=showCheckBoxInRow<br>' +
       '  showDeleteButtonInRow=showDeleteButtonInRow<br>' +
@@ -342,6 +360,7 @@ export default ListFormController.extend({
       '  previousPage=(action "previousPage")<br>' +
       '  gotoPage=(action "gotoPage")<br>' +
       '  nextPage=(action "nextPage")<br>' +
+      '  fixedHeader=fixedHeader<br>' +
       '}}'));
   },
 
@@ -477,6 +496,12 @@ export default ListFormController.extend({
       bindedControllerPropertieName: 'refreshButton'
     });
     componentSettingsMetadata.pushObject({
+      settingName: 'defaultSortingButton',
+      settingType: 'boolean',
+      settingDefaultValue: true,
+      bindedControllerPropertieName: 'defaultSortingButton'
+    });
+    componentSettingsMetadata.pushObject({
       settingName: 'showCheckBoxInRow',
       settingType: 'boolean',
       settingDefaultValue: false,
@@ -535,6 +560,12 @@ export default ListFormController.extend({
       settingType: 'boolean',
       settingDefaultValue: false,
       bindedControllerPropertieName: 'availableCollExpandMode'
+    });
+    componentSettingsMetadata.pushObject({
+      settingName: 'fixedHeader',
+      settingType: 'boolean',
+      settingDefaultValue: false,
+      bindedControllerPropertieName: 'fixedHeader'
     });
 
     return componentSettingsMetadata;