+++ tests/dummy/app/models/ember-flexberry-dummy-suggestion.js
@@ -145,41 +145,29 @@ Model.defineProjection('SuggestionMainModelProjectionTest', 'ember-flexberry-dum
 
 // List form projection.
 Model.defineProjection('SuggestionL', 'ember-flexberry-dummy-suggestion', {
-  address: attr('Address'),
-  text: attr('Text'),
-  date: attr('Date'),
-  votes: attr('Votes'),
-  moderated: attr('Moderated'),
+  address: attr('Address', { index: 0 }),
+  text: attr('Text', { index: 1 }),
+  date: attr('Date', { index: 2 }),
+  votes: attr('Votes', { index: 3 }),
+  moderated: attr('Moderated', { index: 4 }),
   type: belongsTo('ember-flexberry-dummy-suggestion-type', 'Type', {
-    name: attr('Name', {
-      hidden: true
-    })
-  }, {
-    displayMemberPath: 'name'
-  }),
+    name: attr('Name', { index: 6, hidden: true })
+  }, { index: 5, displayMemberPath: 'name' }),
   author: belongsTo('ember-flexberry-dummy-application-user', 'Author', {
-    name: attr('Name', {
-      hidden: true
-    }),
-    eMail: attr('Email')
-  }, {
-    displayMemberPath: 'name'
-  }),
+    name: attr('Name', { index: 8, hidden: true }),
+    eMail: attr('Email', { index: 9 })
+  }, { index: 7, displayMemberPath: 'name' }),
   editor1: belongsTo('ember-flexberry-dummy-application-user', 'Editor', {
-    name: attr('Name', {
-      hidden: true
-    })
-  }, {
-    displayMemberPath: 'name'
-  }),
-  commentsCount: attr('Comments Count'),
+    name: attr('Name', { index: 11, hidden: true })
+  }, { index: 10, displayMemberPath: 'name' }),
+  commentsCount: attr('Comments Count', { index: 15 }),
   comments: hasMany('ember-flexberry-dummy-comment', 'Comments', {
-    text: attr('Text'),
-    votes: attr('Votes'),
-    moderated: attr('Moderated'),
+    text: attr('Text', { index: 0 }),
+    votes: attr('Votes', { index: 1 }),
+    moderated: attr('Moderated', { index: 2 }),
     author: belongsTo('ember-flexberry-dummy-application-user', 'Author', {
-      name: attr('Name', { hidden: true })
-    }, { displayMemberPath: 'name' })
+      name: attr('Name', { index: 4, hidden: true })
+    }, { index: 3, displayMemberPath: 'name' })
   })
 });